requirements-test.txt
Requests==2.25.1
pytest==6.2.3
mock==4.0.3
Requests==2.25.1
kafka-python==2.0.2
psycopg2-binary==2.8.6